backend/contact/views.py
from django.shortcuts import render
from rest_framework import generics
from .models import Contact
from .serializers import ContactSerializer
import os
from dotenv import load_dotenv
from django.http import JsonResponse
from django.views.decorators.csrf import csrf_exempt
from django.core.mail import send_mail
import json

load_dotenv()

@csrf_exempt
def contact_view(request):
    if request.method == 'POST':
        try:
            data = json.loads(request.body)

            name = data.get('name')
            email = data.get('email')
            message = data.get('message')

            subject = f"New Contact Form Submission from {name}"
            message_body = f"Name: {name}\nEmail: {email}\n\nMessage:\n{message}"

            send_mail(
                subject,
                message_body,
                email,  # from
                [os.getenv('EMAIL_HOST_USER')],  # to
                fail_silently=False,
            )

            return JsonResponse({'status': 'success'}, status=200)
        except Exception as e:
            print('Error:', e)
            return JsonResponse({'status': 'error', 'message': str(e)}, status=500)

    return JsonResponse({'error': 'Invalid request'}, status=400)

class ContactListCreate(generics.ListCreateAPIView):
    queryset = Contact.objects.all()
    serializer_class = ContactSerializer
    
class ContactDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Contact.objects.all()
    serializer_class = ContactSerializer
