backend/requirement.txt
Django
django-cors-headers
djangorestframework
sqlparse
psycopg2-binary
python-dotenv
